mcp_neo4j_integration/pyproject.toml
[project]
name = "mcp-neo4j-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.8",
    "python-dotenv>=1.1.0",
]
